requirements.txt
browser_use==0.1.35
langchain_openai==0.3.3
python-dotenv==1.0.1
rich==13.9.4
setuptools==75.8.0
